+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   * Bump ammonia to 4.0
   * Add new `extension` options: multiline_block_quotes, math_dollars, math_code, shortcodes, wikilinks_title_after_pipe, wikilinks_title_before_pipe
   * Add new `render` option: escaped_char_spans
+  * Add new option `features.syntax_highlight_inline_style` to control whether to embed inline styles or not. Default is `true`.
 
 ## 0.1.16 (2024-04-29)